+++ packages/adminui-pub/_src/app/core/UBUtilTree.js
@@ -1,64 +1,52 @@
-﻿/**
- * Файл: UB.core.UBUtilTree.js
- * Автор: Игорь Ноженко
- *
- * Функции для формирования данных для tree
+/*
+ * author: Nozhenko
  */
+UB.core.UBUtilTree = {
+  arrayToTreeRootNode,
+  addToParent
+}
 
-Ext.define("UB.core.UBUtilTree", {
-    singleton: true,
+function arrayToTreeRootNode (data) {
+  let root = {
+    leaf: false,
+    id: 0,
+    expanded: true
+  }
 
-    pathSeparator: "/",
-    pathAttribute: "treepath",
-    doSort: false,
+  let items = { 0: root }
 
-    arrayToTreeRootNode: function (data) {
-        var
-            root = {
-                leaf: false,
-                id: 0,
-                expanded: true
-            },
-            items = {0: root}, item, parent,
-            i,
-            len;
+  for (let i = 0, len = data.length; i < len; ++i) {
+    let row = data[i]
 
-        for (i = 0, len = data.length; i < len; ++i) {
-            var
-                row = data[i];
-
-            item = items[row.id];
-            if (!item){
-                item = items[row.id] = row;
-            } else {
-                Ext.applyIf(item, row);
-            }
-            parent = items[row.parentId || 0];
-            if (!parent){
-                parent = items[row.parentId] = {id: row.parentId};
-            }
-            if (!parent.children){
-                parent.children = [];
-            }
-            parent.children.push(item);
-        }
-
-        return root;
-    },
-
-    /**
-     *
-     * @param {Ext.data.NodeInterface} parent
-     * @param {Object[]} childNodes
-     */
-    addToParent: function(parent, childNodes){
-        _.forEach(childNodes, function(node){
-             var elm = parent.appendChild(node);
-             if (node.children && node.children.length > 0){
-                 UB.core.UBUtilTree.addToParent(elm, node.children);
-             }
-
-        })
+    let item = items[row.id]
+    if (!item) {
+      item = items[row.id] = row
+    } else {
+      Ext.applyIf(item, row)
     }
+    let parent = items[row.parentId || 0]
+    if (!parent) {
+      parent = items[row.parentId] = { id: row.parentId }
+    }
+    if (!parent.children) {
+      parent.children = []
+    }
+    parent.children.push(item)
+  }
 
-});
+  return root
+}
+
+/**
+ *
+ * @param {Ext.data.NodeInterface} parent
+ * @param {Array<object>} childNodes
+ */
+function addToParent (parent, childNodes) {
+  childNodes.forEach(function (node) {
+    let elm = parent.appendChild(node)
+    if (node.children && node.children.length) {
+      UB.core.UBUtilTree.addToParent(elm, node.children)
+    }
+  })
+}